Add clean script to remove test-browser.js after the tests run
Whenever you run `npm test` it would generate a file.
The location of the file would be `tests/browser/test-browser.js`
which is a directory that the linter will lint.
I tried making the linter ignore the file but it seems we're using
an older version of the linter which may not support ignoring paths.

I may be wrong but this was an alternative solution, safely delete
the file before running the linter.