Pause/Resume response content, not response
When the response content is piped through additional streams for
decoding (e.g. for gzip decompression), pause and resume calls should be
propagated to the last stream in the pipeline so that back pressure
propagates correctly.

This avoids an issue where simultaneous back pressure from the content
decoding stream and from a stream to which Request is piped could cause
the response stream to get stuck waiting for a drain event on the
content decoding stream which never occurs.  See #1567 for an example.

This commit also renames dataStream to responseContent to remedy my
previous poor choice of name, since the name will be exposed on the
Request instance it should be clearer and closer to the name used to
refer to this data in the relevant RFCs.

Fixes #1567

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>