More lenient gzip decompression
Be explicitly lenient with gzip decompression by always requesting
`zlib` to flush the input data and never explicitly ending the
`zlib` input.

The behavioural difference is that on Node ≥ 6, which has a slightly
stricter gzip decoding process than previous Node versions, malformed
but otherwise acceptable server responses are still properly
decompressed (the most common example being a missing checksum
at the stream end).

This aligns behaviour with cURL, which always uses the `Z_SYNC_FLUSH`
flag for decompression.

On the downside, accidental truncation of a response is no longer
detected on the compression layer.

Ref: https://github.com/nodejs/node/issues/8701#issuecomment-268224481
Fixes: https://github.com/request/request/issues/2482